requirements.txt
clarifai-grpc==9.10.0
clarifai==9.10.1
scikit-learn==1.3.2
umap-learn==0.5.2
streamlit==1.27.0
requests==2.28.1
numpy==1.26.0
plotly==5.11.0
pandas==2.1.0
stqdm==0.0.5